// src/app/models/rnInfo.model.ts
export interface rnInfo {
	//demoid: string;
    orgId: string;
	rnId: string;
	firstName: string;
	lastName: string;
	Organizationemailaddress: string;
	cohortDate: string;
	employmentStartDate: string;
	typeOfUnit: number;
	typeOfUnitOther: string;
	age: number;
	gender: number;
	ethnicity: number;
	education: string;
	degreeReceived: string;
	degreeReceivedOther: string;
	previousNursingDegree: string;
	major: string;
	overallGPA: number;	
	meetingPlaceAccess: string;
	//lastUpdatedBy: string;
	
	//"Participation TYPE": string;
	
	//unittypedesc: string;
	//ABSENT: string;
	
}